Move HabitList back to Components because it actualy is a component
